# Insight-Performance-Hub/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Organisational Performance Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sdbip-field-configs:
    get:
      operationId: listSdbipFieldConfigs
      tags: [sdbipFieldConfigs]
      summary: List SDBIP field configurations
      parameters:
        - name: sdbipType
          in: query
          required: false
          schema:
            type: string
            enum: [original, revised, departmental, quarterly, midyear, annual]
      responses:
        "200":
          description: List of field configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SdbipFieldConfig"
  /sdbip-field-configs/{sdbipType}:
    put:
      operationId: saveSdbipFieldConfigs
      tags: [sdbipFieldConfigs]
      summary: Replace field configuration for an SDBIP type
      parameters:
        - name: sdbipType
          in: path
          required: true
          schema:
            type: string
            enum: [original, revised, departmental, quarterly, midyear, annual]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveSdbipFieldConfigsInput"
      responses:
        "200":
          description: Saved field configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SdbipFieldConfig"
  /sdbip-field-configs/{sdbipType}/usage:
    get:
      operationId: getSdbipFieldUsage
      tags: [sdbipFieldConfigs]
      summary: Count KPIs holding data for each configured field
      parameters:
        - name: sdbipType
          in: path
          required: true
          schema:
            type: string
            enum: [original, revised, departmental]
      responses:
        "200":
          description: KPI data counts keyed by fieldKind:fieldKey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SdbipFieldUsage"
...
components:
  schemas:
...
    SdbipFieldConfig:
      type: object
      properties:
        id:
          type: integer
        sdbipType:
          type: string
          enum: [original, revised, departmental, quarterly, midyear, annual]
        fieldKind:
          type: string
          enum: [primary, custom]
        fieldKey:
          type: string
        fieldLabel:
          type: string
        fieldType:
          type: string
          enum: [text, number, date, boolean, textarea, select, alphanumeric, percent]
        isIncluded:
          type: boolean
        isRequired:
          type: boolean
        isLocked:
          type: boolean
        sortOrder:
          type: integer
      required:
        - id
        - sdbipType
        - fieldKind
        - fieldKey
        - fieldLabel
        - fieldType
        - isIncluded
        - isRequired
        - isLocked
        - sortOrder
    SdbipFieldConfigItemInput:
      type: object
      properties:
        fieldKind:
          type: string
          enum: [primary, custom]
        fieldKey:
          type: string
        fieldLabel:
          type: string
        fieldType:
          type: string
          enum: [text, number, date, boolean, textarea, select, alphanumeric, percent]
          default: text
        isIncluded:
          type: boolean
          default: true
        isRequired:
          type: boolean
          default: false
        isLocked:
          type: boolean
          default: false
        sortOrder:
          type: integer
          default: 0
      required:
        - fieldKind
        - fieldKey
        - fieldLabel
    SaveSdbipFieldConfigsInput:
      type: object
      properties:
        fields:
          type: array
          items:
            $ref: "#/components/schemas/SdbipFieldConfigItemInput"
      required:
        - fields
    SdbipFieldUsage:
      type: object
      properties:
        usage:
          type: object
          additionalProperties:
            type: integer
      required:
        - usage
...